+++ facade/src/test/scala/react/semanticui/collections/form/FormInputTests.scala
@@ -5,10 +5,10 @@ import japgolly.scalajs.react.test._
 import japgolly.scalajs.react.vdom.html_<^._
 import react.semanticui.widths._
 
-object FormCheckboxTests extends TestSuite {
+object FormInputTests extends TestSuite {
   val tests = Tests {
     test("render") {
-      val form = FormCheckbox()
+      val form = FormInput()
       ReactTestUtils.withNewBodyElement { mountNode =>
         form.renderIntoDOM(mountNode)
         assert(
@@ -17,20 +17,21 @@ object FormCheckboxTests extends TestSuite {
       }
     }
     test("width") {
-      val form = FormCheckbox(width = Two)
+      val form = FormInput(width = Two)
       ReactTestUtils.withNewBodyElement { mountNode =>
         form.renderIntoDOM(mountNode)
+        println(mountNode.innerHTML)
         assert(
-          mountNode.innerHTML == """<div class="two wide field"><div class="ui fitted checkbox"><input class="hidden" readonly="" tabindex="0" type="checkbox" value=""><label></label></div></div>"""
+          mountNode.innerHTML == """<div class="two wide field"><div class="ui input"><input type="text"></div></div>"""
         )
       }
     }
-    test("value") {
-      val form = FormCheckbox(width = Two, value = 1, slider = true)
+    test("inverted") {
+      val form = FormInput(width = Two, inverted = true, focus = true, value = "test")
       ReactTestUtils.withNewBodyElement { mountNode =>
         form.renderIntoDOM(mountNode)
         assert(
-          mountNode.innerHTML == """<div class="two wide field"><div class="ui fitted slider checkbox"><input class="hidden" readonly="" tabindex="0" type="checkbox" value="1"><label></label></div></div>"""
+          mountNode.innerHTML == """<div class="two wide field"><div class="ui focus inverted input"><input type="text" value="test"></div></div>"""
         )
       }
     }